Move window-based aggregates to contrib
