+++ contrib/src/main/java/macrobase/analysis/transform/SlidingWindowTransform.java
