+++ contrib/src/main/java/macrobase/analysis/transform/aggregate/BatchAggregate.java
