+++ contrib/src/main/java/macrobase/analysis/transform/aggregate/BatchWindowAggregate.java
