+++ contrib/src/main/java/macrobase/analysis/transform/aggregate/BatchWindowMax.java
