+++ contrib/src/main/java/macrobase/analysis/transform/aggregate/IncrementalAggregate.java
