+++ contrib/src/main/java/macrobase/analysis/transform/aggregate/IncrementalWindowAggregate.java
