+++ contrib/src/main/java/macrobase/analysis/transform/aggregate/IncrementalWindowCount.java
