+++ contrib/src/main/java/macrobase/analysis/transform/aggregate/IncrementalWindowSum.java
