+++ contrib/src/main/java/macrobase/analysis/transform/aggregate/PaneAggregate.java
