+++ contrib/src/test/java/macrobase/analysis/stats/MovingAverageTest.java
