+++ contrib/src/test/java/macrobase/analysis/transform/aggregate/BatchWindowMaxTest.java
