+++ contrib/src/test/java/macrobase/analysis/transform/aggregate/IncrementalWindowCountTest.java
