+++ core/src/main/java/macrobase/conf/MacroBaseDefaults.java
@@ -49,7 +49,6 @@ public class MacroBaseDefaults {
     public static final Integer TIME_COLUMN = null;
     public static final MacroBaseConf.TransformType TRANSFORM_TYPE = MacroBaseConf.TransformType.MAD_OR_MCD;
     public static final CSVIngester.Compression CSV_COMPRESSION = CSVIngester.Compression.UNCOMPRESSED;
-    public static final MacroBaseConf.AggregateType AGGREGATE_TYPE = MacroBaseConf.AggregateType.COUNT;
 
     public static final String DB_USER = System.getProperty("user.name");
     public static final String DB_PASSWORD = "";
